requirements.txt
torch
transformers
datasets
pandas
scikit-learn
openai
python-dotenv
asyncio